autotools: fix static linking with pkg-config
The CMake scripts for libzen and libmediainfo don't use Libs_Static
so LIBS would always be empty. Additionally, the proper way to get
static libcurl libs is to use '--static --libs' when invoking
pkg-config.

To prevent regressions, do both checks for Libs_Static variable
and the new '--static --libs' and use both. It's kind of overlinking,
maybe, but should prevent regressions for now.